% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_num.R
\name{convert_to_num}
\alias{convert_to_num}
\title{Converts columns that can be coerced to numeric type in a
data frame to numeric type.}
\usage{
convert_to_num(df, cols)
}
\arguments{
\item{df}{The dataframe containing columns to be converted}

\item{cols}{A vector of column names to be converted to numeric type}
}
\value{
A dataframe containing converted numeric columns
}
\description{
Converts columns that can be coerced to numeric type in a
data frame to numeric type.
}
\examples{
convert_to_num(data.frame(c_to_n = c('1','2')), 'c_to_n')
}
