% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_load.R
\name{df_load}
\alias{df_load}
\title{This function load and preprocess a dataset from a CSV file,
split the date frame into 2 parts that the first part is nice
formatted and the second part need manipulation,
correct the wrong row entries in the second part,
convert the numeric columns into correct data type,
combine the data together.}
\usage{
df_load(
  url,
  skip1,
  skip2,
  n_max1,
  n_max2,
  error_line,
  error_record,
  correct_bef_error_record,
  val_corrected,
  error_col,
  predicted_factor
)
}
\arguments{
\item{url}{The quoted URL or file path of the CSV file.}

\item{skip1}{The number of rows to skip at the beginning
of the file for the first splited dataset.}

\item{skip2}{The number of rows to skip at the beginning
of the file for the second splited dataframe.}

\item{n_max1}{The maximum number of rows to read for the
first splited dataframe.}

\item{n_max2}{The maximum number of rows to read for the
second splited dataframe.}

\item{error_line}{The row index containing errors that
need to be corrected.}

\item{error_record}{A vector of the column indices containing
errors that need to be corrected.}

\item{correct_bef_error_record}{The number of correctly formatted
rows before the row containing errors.}

\item{val_corrected}{list of corrected entries to the error_line.}

\item{error_col}{A vector of the corrected values for the error columns.}

\item{predicted_factor}{The quoted column name to be converted to a factor.}
}
\value{
The preprocessed data frame, with missing values corrected and data
types converted.
}
\description{
This function load and preprocess a dataset from a CSV file,
split the date frame into 2 parts that the first part is nice
formatted and the second part need manipulation,
correct the wrong row entries in the second part,
convert the numeric columns into correct data type,
combine the data together.
}
\examples{
df_load(
  url = "https://raw.githubusercontent.com/mwaskom/seaborn-data/master/iris.csv",
  skip1 = 0,
  skip2 = 0,
  n_max1 = 3,
  n_max2 = 2,
  error_line = 2,
  error_record = 2:4,
  correct_bef_error_record = 1,
  val_corrected = list(3.5, 1.4, 0.2, "setosa"),
  error_col = c("petal_width"),
  predicted_factor = "species"
)
}
