% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data function}
\usage{
filter_data(data, col_to_filter_by, subset, start_col, end_col, alg)
}
\arguments{
\item{data}{Data to be filtered}

\item{col_to_filter_by}{Name of the column which should be used
for filtering the data.}

\item{subset}{The filtering condition which the targeted function
should follow.}

\item{start_col}{The start of the column range which should be
selected from the data.}

\item{end_col}{The end of the column range which should be
selected from the data.}

\item{alg}{The filtering algorithm which the targeted function
should follow.}
}
\value{
Return the data filtered by the parameters above.
}
\description{
This function calculates metrics based on a given algorithm
(e.g. mean, range, etc.) on the used dataset (e.g. fire_train)
filters the data by a given column name and a value and by a
column range
}
\examples{
filter_data(
  data.frame(
    type = c("big", "big", "small"),
    length = c(5.2, 2.1, 1.2),
    height = c(10.1, 20.1, 4.2),
    depth = c(3, 2, 1.8)
  ), type, "big", height, depth, mean
)
}
