% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn.R
\name{plot_line_graph}
\alias{plot_line_graph}
\title{This function plots a line graph given the data and the column names to
visualize and enables to customize the graphs axis labels and plot dimensions}
\usage{
plot_line_graph(
  data,
  plot_width,
  plot_height,
  x_axis_data,
  y_axis_data,
  x_axis_label,
  y_axis_label
)
}
\arguments{
\item{data}{The data which should be visualized}

\item{plot_width}{The width of the plot}

\item{plot_height}{The height of the plot}

\item{x_axis_data}{The data which should be displayed on the x axis}

\item{y_axis_data}{The data which should be displayed on the y axis}

\item{x_axis_label}{Label which will be displayed in the plot for the x axis}

\item{y_axis_label}{Label which will be displayed in the plot for the y axis}
}
\value{
The generated plot
}
\description{
This function plots a line graph given the data and the column names to
visualize and enables to customize the graphs axis labels and plot dimensions
}
\examples{
plot_line_graph(
  data = data.frame(neighbors = c(1, 2, 3), mean = c(0.80, 0.85, 0.95)),
  plot_width = 10, plot_height = 10,
  x_axis_data = neighbors, y_axis_data = mean, x_axis_label = "Neighbors",
  y_axis_label = "Accuracy Estimate"
)
}
