% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn.R
\name{plot_scatter_graph}
\alias{plot_scatter_graph}
\title{This function plots a scatter graph given the data and the column names to
visualize and enables to customize the graphs axis labels and plot dimensions}
\usage{
plot_scatter_graph(
  data,
  plot_width,
  plot_height,
  x_axis_data,
  y_axis_data,
  x_axis_label,
  y_axis_label,
  text_size,
  color,
  color_label
)
}
\arguments{
\item{data}{The data which should be visualized}

\item{plot_width}{The width of the plot}

\item{plot_height}{The height of the plot}

\item{x_axis_data}{The data which should be displayed on the x axis}

\item{y_axis_data}{The data which should be displayed on the y axis}

\item{x_axis_label}{Label which will be displayed in the plot for the x axis}

\item{y_axis_label}{Label which will be displayed in the plot for the y axis}

\item{text_size}{Text size applied for all text in the plot}

\item{color}{Name of the label column of the dataset}

\item{color_label}{Defines labels for each of the colors}
}
\value{
The generated plot
}
\description{
This function plots a scatter graph given the data and the column names to
visualize and enables to customize the graphs axis labels and plot dimensions
}
\examples{
plot_scatter_graph(
  data = data.frame(
    ISI = c(5.2, 2.1, 5.2), BUI = c(10.1, 2.1, 4.2),
    Classes = c("fire", "fire", "no fire")
  ),
  plot_width = 10, plot_height = 10, x_axis_data = ISI,
  y_axis_data = BUI, x_axis_label = "ISI",
  y_axis_label = "BUI", text_size = 20, color = Classes,
  color_label = "Presence of fire"
)
}
